Corrected error in identifying unique rows Added functionality to map each event to specific program stage
