+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "importer",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "private": true,
   "dependencies": {
     "@babel/core": "7.1.0",
@@ -78,7 +78,7 @@
     "terser-webpack-plugin": "1.1.0",
     "url-loader": "1.1.1",
     "webpack": "4.19.1",
-    "webpack-dev-server": "3.1.9",
+    "webpack-dev-server": "^3.3.1",
     "webpack-manifest-plugin": "2.0.4",
     "webworkify-webpack": "^2.1.3",
     "workbox-webpack-plugin": "3.6.3",
