+++ src/components/program/step5.js
@@ -3,6 +3,7 @@ import {withStyles} from "@material-ui/core/styles";
 
 import {inject, observer} from "mobx-react";
 import Summary from "./Summary";
+import {NotificationManager} from "react-notifications";
 
 const styles = theme => ({
     margin: {
@@ -23,6 +24,14 @@ class Step5 extends React.Component {
         super(props);
         const {IntegrationStore} = props;
         this.integrationStore = IntegrationStore;
+    }
+
+    componentDidMount() {
+        const imports = this.integrationStore.program.totalImports;
+
+        if (imports === 0) {
+            NotificationManager.info(`Importer could not find what to import or update, records might be upto date. Click next to save mapping`, 'Information', 10000);
+        }
     }
 
     render() {
