+++ src/stores/DataSet.js
@@ -1,9 +1,9 @@
 import {action, computed, observable} from "mobx";
 import _ from 'lodash';
 import XLSX from "xlsx";
-import {encodeData, findAttributeCombo, nest, processMergedCells, enumerateDates} from "../utils";
+import {encodeData, nest, processMergedCells, enumerateDates, processDataSet} from "../utils";
 import axios from "axios";
-import alasql from 'alasql';
+// import alasql from 'alasql';
 import {NotificationManager} from "react-notifications";
 import Param from "./Param";
 
@@ -1044,235 +1044,8 @@ class DataSet {
     }
 
     @computed get processed() {
-        let data = this.data;
-        let dataValues = [];
 
-        let dataSetUnits;
-
-        if (this.orgUnitStrategy) {
-            dataSetUnits = _.fromPairs(this.organisationUnits.map(o => {
-                if (this.orgUnitStrategy && this.orgUnitStrategy.value === 'name') {
-                    return [o.name.toLowerCase(), o.id];
-                } else if (this.orgUnitStrategy && this.orgUnitStrategy.value === 'code') {
-                    return [o.code, o.id];
-                }
-                return [o.id, o.id];
-
-            }));
-        }
-        if (this.templateType !== '4') {
-            const forms = this.forms;
-            forms.forEach(f => {
-                if (this.templateType === '1') {
-
-                    let validatedData = [];
-                    f.dataElements.forEach(element => {
-                        if (element.mapping) {
-                            const foundData = data[element.mapping.value];
-                            if (foundData) {
-                                const groupedData = foundData.map(d => {
-
-                                    const rowData = this.categoryCombo.categories.map(category => {
-                                        const optionColumn = category.mapping.value;
-                                        return d[optionColumn]
-                                    });
-
-                                    return {
-                                        period: d[this.periodColumn.value],
-                                        value: d[this.dataValueColumn.value],
-                                        orgUnit: d[this.orgUnitColumn.value] ? d[this.orgUnitColumn.value].toLocaleLowerCase() : null,
-                                        dataElement: element.id,
-                                        attributeValue: rowData,
-                                        categoryOptionCombo: d[this.categoryOptionComboColumn.value] ? d[this.categoryOptionComboColumn.value].toLocaleLowerCase() : null
-                                    }
-                                });
-                                validatedData = [...validatedData, ...groupedData];
-                            }
-                        }
-                    });
-                    data = validatedData;
-                }
-                if (data) {
-                    f.categoryOptionCombos.forEach(coc => {
-                        if (this.templateType === '1') {
-                            // f.categoryOptionCombos.forEach(coc => {
-                            _.forOwn(coc.mapping, (mapping, dataElement) => {
-                                const filtered = data.filter(v => {
-                                    return mapping && mapping.value && v.categoryOptionCombo === mapping.value.toLocaleLowerCase() && v.dataElement === dataElement;
-                                });
-                                filtered.forEach(d => {
-                                    const attribute = findAttributeCombo(this, d.attributeValue, false);
-                                    if (d['orgUnit'] && attribute) {
-                                        const orgUnit = dataSetUnits[d['orgUnit']];
-                                        if (orgUnit) {
-                                            dataValues = [...dataValues, {
-                                                dataElement,
-                                                value: d['value'],
-                                                period: d['period'],
-                                                attributeOptionCombo: attribute.id,
-                                                categoryOptionCombo: coc.id,
-                                                orgUnit
-                                            }];
-                                        }
-                                    }
-                                });
-                            });
-                            // });
-                        } else if (this.templateType === '2') {
-                            _.forOwn(coc.cell, (mapping, dataElement) => {
-                                let orgUnit;
-                                let period;
-                                if (!this.periodInExcel) {
-                                    period = this.period;
-                                } else if (this.periodColumn) {
-                                    const p = data[this.periodColumn.value]['v'];
-                                    period = p.toString();
-                                }
-
-
-                                if (!this.organisationUnitInExcel) {
-                                    orgUnit = this.organisation.value
-                                } else {
-                                    const ou = data[this.organisationCell.value]['v'];
-                                    const foundOU = dataSetUnits[ou];
-                                    if (foundOU) {
-                                        orgUnit = foundOU;
-                                    } else {
-                                        NotificationManager.error(`Organisation unit ${ou} not found`);
-                                    }
-                                }
-
-                                let found;
-
-                                if (this.attributeCombosInExcel) {
-                                    const rowData = this.categoryCombo.categories.map(category => {
-                                        const value = data[category.mapping.value];
-                                        return value ? value.v : undefined;
-                                    });
-                                    found = findAttributeCombo(this, rowData, false);
-
-                                } else {
-                                    const rowData = this.categoryCombo.categories.map(category => {
-                                        return category.mapping.value;
-                                    });
-                                    found = findAttributeCombo(this, rowData, true);
-                                }
-                                if (found) {
-                                    dataValues = [...dataValues, {
-                                        dataElement,
-                                        value: data[mapping.value]['v'],
-                                        categoryOptionCombo: coc.id,
-                                        period,
-                                        attributeOptionCombo: found.id,
-                                        orgUnit
-                                    }]
-                                }
-
-
-                            });
-                        } else if (this.templateType === '3') {
-                            if (this.rows) {
-                                this.rows.forEach(r => {
-                                    const rowData = this.categoryCombo.categories.map(category => {
-                                        const optionCell = category.mapping.value + r;
-                                        const optionValue = this.data[optionCell];
-                                        return optionValue ? optionValue.v : undefined;
-                                    });
-                                    const found = findAttributeCombo(this, rowData, false);
-                                    if (found) {
-                                        _.forOwn(coc.column, (mapping, dataElement) => {
-                                            const cell = mapping.value + r;
-                                            let orgUnit = data[this.organisationColumn.value + r]['v'];
-                                            let period = data[this.periodColumn.value + r]['v'];
-                                            let value = data[cell]['v'];
-                                            orgUnit = dataSetUnits[orgUnit];
-                                            dataValues = [...dataValues, {
-                                                dataElement,
-                                                value,
-                                                categoryOptionCombo: coc.id,
-                                                attributeOptionCombo: found.id,
-                                                period,
-                                                orgUnit
-                                            }]
-                                        });
-                                    }
-                                })
-                            }
-                        }
-                    });
-                }
-            });
-        } else if (this.templateType === '4') {
-            let periodMissing = false;
-            let valueMissing = false;
-            let orgUnitMissing = false;
-            this.rows.forEach(i => {
-                const rowData = this.categoryCombo.categories.map(category => {
-                    const optionCell = category.mapping.value + i;
-                    const optionValue = this.data[optionCell];
-                    return optionValue ? optionValue.v : undefined;
-                });
-
-                const found = findAttributeCombo(this, rowData, false);
-                if (found) {
-                    _.forOwn(this.cell2, v => {
-                        const oCell = this.orgUnitColumn.value + i;
-                        const pCell = this.periodColumn.value + i;
-                        const vCell = v.value.column + i;
-                        const ouVal = this.data[oCell];
-                        const periodVal = this.data[pCell];
-                        const ou = ouVal ? ouVal['v'].toLowerCase() : '';
-                        const period = periodVal ? periodVal['v'] : null;
-                        const val = this.data[vCell];
-                        const value = val ? val.v : null;
-
-                        const orgUnit = dataSetUnits[ou];
-                        if (orgUnit && value && period) {
-                            dataValues = [...dataValues, {
-                                orgUnit,
-                                period,
-                                value,
-                                dataElement: v.value.dataElement,
-                                attributeOptionCombo: found.id,
-                                categoryOptionCombo: v.value.categoryOptionCombo
-                            }];
-                        } else {
-                            if (!orgUnit) {
-                                orgUnitMissing = true;
-                            }
-                            if (!period) {
-                                periodMissing = true;
-                            }
-
-                            if (!value) {
-                                valueMissing = true;
-                            }
-                        }
-
-                    });
-                }
-            });
-
-            if (orgUnitMissing) {
-                NotificationManager.warning(`Some rows are missing organisation units, will be ignored`);
-            }
-            if (periodMissing) {
-                NotificationManager.warning(`Some rows are missing periods, will be ignored`);
-            }
-
-            if (valueMissing) {
-                NotificationManager.warning(`Some rows are missing values, will be ignored`);
-            }
-        }
-        dataValues = dataValues.filter(dv => {
-            return dv.orgUnit && dv.period
-        });
-        if (dataValues.length > 0) {
-            return alasql('SELECT orgUnit,dataElement,attributeOptionCombo,categoryOptionCombo,period,SUM(`value`) AS `value` FROM ? GROUP BY orgUnit,dataElement,attributeOptionCombo,categoryOptionCombo,period', [dataValues]);
-        }
-
-
-        return dataValues
+        return processDataSet(this.data, this)
 
     }
