+++ src/utils.js
@@ -19,6 +19,9 @@ import Program from "./stores/Program";
 import TrackedEntityType from "./stores/TrackedEntityType";
 import Param from "./stores/Param";
 import moment from "moment";
+import {generateUid} from "d2/uid";
+// import {NotificationManager} from "react-notifications";
+// import alasql from "alasql";
 
 export const nest = function (seq, keys) {
     if (!keys.length)
@@ -182,6 +185,7 @@ export const convertAggregate = (ds, d2) => {
 
         f.setCategoryOptionCombos(categoryOptionCombos);
         f.setDataElements(dataElements);
+        f.setName(form.name);
 
         return f;
 
@@ -282,11 +286,6 @@ export const convert = (program, d2) => {
             const programStageDataElement = new ProgramStageDataElement(psd.compulsory, dataElement);
             if (psd.column) {
                 programStageDataElement.setColumn(psd.column);
-            } else {
-                programStageDataElement.setColumn({
-                    name: null,
-                    value: null
-                });
             }
             programStageDataElements = [...programStageDataElements, programStageDataElement];
         });
@@ -301,6 +300,9 @@ export const convert = (program, d2) => {
         programsStage.setCompleteEvents(ps.completeEvents);
         programsStage.setLongitudeColumn(ps.longitudeColumn);
         programsStage.setLatitudeColumn(ps.latitudeColumn);
+        programsStage.setCreateNewEvents(ps.createNewEvents);
+        programsStage.setUpdateEvents(ps.updateEvents);
+        programsStage.setEventDateColumn(ps.eventDateColumn);
         programStages = [...programStages, programsStage]
     });
 
@@ -357,17 +359,14 @@ export const convert = (program, d2) => {
     }
 
     p.setD2(d2);
-    p.setOrder(program.order);
-    p.setOrderBy(program.orderBy);
+    // p.setOrder(program.order);
+    // p.setOrderBy(program.orderBy);
     p.setOrgUnitStrategy(program.orgUnitStrategy);
     p.setHeaderRow(program.headerRow || 1);
     p.setDataStartRow(program.dataStartRow || 2);
-    p.setCreateNewEvents(program.createNewEvents);
-    p.setUpdateEvents(program.updateEvents);
     p.setCreateNewEnrollments(program.createNewEnrollments);
     p.setCreateEntities(program.createEntities);
     p.setUpdateEntities(program.updateEntities);
-    p.setEventDateColumn(program.eventDateColumn);
     p.setEnrollmentDateColumn(program.enrollmentDateColumn);
     p.setIncidentDateColumn(program.incidentDateColumn);
     p.setUrl(program.url || '');
@@ -405,17 +404,6 @@ export const findAttributeCombo = (dataSet, data, compareId) => {
     });
 };
 
-// export const searchTrackedEntityInstance = async (d2, uniqueAttribute, value) => {
-//     const api = d2.Api.getApi();
-//
-//     const instance = await api.get('trackedEntityInstances', {
-//         paging: false,
-//         ouMode: 'ALL',
-//         filter: uniqueAttribute + ':EQ:' + value,
-//         fields: 'trackedEntityInstance'
-//     });
-// };
-
 
 export const encodeData = (objs) => {
     return objs.map(s => {
@@ -441,4 +429,995 @@ export const enumerateDates = (startDate, endDate, addition, format) => {
         dates.push(currDate.clone().format(format));
     }
     return dates;
+};
+
+export const processOrganisationUnits = (dataSet) => {
+    if (dataSet.orgUnitStrategy && dataSet.organisationUnits) {
+        return _.fromPairs(dataSet.organisationUnits.map(o => {
+            if (dataSet.orgUnitStrategy && dataSet.orgUnitStrategy.value === 'name') {
+                return [o.name.toLowerCase(), o.id];
+            } else if (dataSet.orgUnitStrategy && dataSet.orgUnitStrategy.value === 'code') {
+                return [o.code, o.id];
+            }
+            return [o.id, o.id];
+
+        }));
+    }
+    return {};
+};
+
+
+export const processDataSet = (data, dataSet) => {
+    let dataValues = [];
+
+    let dataSetUnits = processOrganisationUnits(dataSet);
+
+    const {
+        templateType,
+        forms,
+        periodColumn,
+        dataValueColumn,
+        orgUnitColumn,
+        categoryCombo,
+        categoryOptionComboColumn,
+        periodInExcel,
+        organisationUnitInExcel,
+        organisation,
+        organisationCell,
+        attributeCombosInExcel,
+        rows,
+        cell2,
+        organisationColumn
+    } = dataSet;
+
+
+    if (templateType !== '4') {
+        forms.forEach(f => {
+            if (templateType === '1') {
+
+                let validatedData = [];
+                f.dataElements.forEach(element => {
+                    if (element.mapping) {
+                        const foundData = data[element.mapping.value];
+                        if (foundData) {
+                            const groupedData = foundData.map(d => {
+
+                                const rowData = categoryCombo.categories.map(category => {
+                                    const optionColumn = category.mapping.value;
+                                    return d[optionColumn]
+                                });
+
+                                return {
+                                    period: d[periodColumn.value],
+                                    value: d[dataValueColumn.value],
+                                    orgUnit: d[orgUnitColumn.value] ? d[orgUnitColumn.value].toLocaleLowerCase() : null,
+                                    dataElement: element.id,
+                                    attributeValue: rowData,
+                                    categoryOptionCombo: d[categoryOptionComboColumn.value] ? d[categoryOptionComboColumn.value].toLocaleLowerCase() : null
+                                }
+                            });
+                            validatedData = [...validatedData, ...groupedData];
+                        }
+                    }
+                });
+                data = validatedData;
+            }
+            if (data) {
+                f.categoryOptionCombos.forEach(coc => {
+                    if (templateType === '1') {
+                        _.forOwn(coc.mapping, (mapping, dataElement) => {
+                            const filtered = data.filter(v => {
+                                return mapping && mapping.value && v.categoryOptionCombo === mapping.value.toLocaleLowerCase() && v.dataElement === dataElement;
+                            });
+                            filtered.forEach(d => {
+                                const attribute = findAttributeCombo(dataSet, d.attributeValue, false);
+                                if (d['orgUnit'] && attribute) {
+                                    const orgUnit = dataSetUnits[d['orgUnit']];
+                                    if (orgUnit) {
+                                        dataValues = [...dataValues, {
+                                            dataElement,
+                                            value: d['value'],
+                                            period: d['period'],
+                                            attributeOptionCombo: attribute.id,
+                                            categoryOptionCombo: coc.id,
+                                            orgUnit
+                                        }];
+                                    }
+                                }
+                            });
+                        });
+                    } else if (templateType === '2') {
+                        _.forOwn(coc.cell, (mapping, dataElement) => {
+                            let orgUnit;
+                            let period;
+                            if (!periodInExcel) {
+                                period = dataSet.period;
+                            } else if (periodColumn) {
+                                const p = data[periodColumn.value]['v'];
+                                period = p.toString();
+                            }
+
+
+                            if (!organisationUnitInExcel) {
+                                orgUnit = organisation.value
+                            } else {
+                                const ou = data[organisationCell.value]['v'];
+                                const foundOU = dataSetUnits[ou];
+                                if (foundOU) {
+                                    orgUnit = foundOU;
+                                } else {
+                                    // NotificationManager.error(`Organisation unit ${ou} not found`);
+                                }
+                            }
+
+                            let found;
+
+                            if (attributeCombosInExcel) {
+                                const rowData = categoryCombo.categories.map(category => {
+                                    const value = data[category.mapping.value];
+                                    return value ? value.v : undefined;
+                                });
+                                found = findAttributeCombo(dataSet, rowData, false);
+
+                            } else {
+                                const rowData = categoryCombo.categories.map(category => {
+                                    return category.mapping.value;
+                                });
+                                found = findAttributeCombo(dataSet, rowData, true);
+                            }
+                            if (found) {
+                                dataValues = [...dataValues, {
+                                    dataElement,
+                                    value: data[mapping.value]['v'],
+                                    categoryOptionCombo: coc.id,
+                                    period,
+                                    attributeOptionCombo: found.id,
+                                    orgUnit
+                                }]
+                            }
+
+
+                        });
+                    } else if (templateType === '3') {
+                        if (rows) {
+                            rows.forEach(r => {
+                                const rowData = categoryCombo.categories.map(category => {
+                                    const optionCell = category.mapping.value + r;
+                                    const optionValue = data[optionCell];
+                                    return optionValue ? optionValue.v : undefined;
+                                });
+                                const found = findAttributeCombo(dataSet, rowData, false);
+                                if (found) {
+                                    _.forOwn(coc.column, (mapping, dataElement) => {
+                                        const cell = mapping.value + r;
+                                        let orgUnit = data[organisationColumn.value + r]['v'];
+                                        let period = data[periodColumn.value + r]['v'];
+                                        let value = data[cell]['v'];
+                                        orgUnit = dataSetUnits[orgUnit];
+                                        dataValues = [...dataValues, {
+                                            dataElement,
+                                            value,
+                                            categoryOptionCombo: coc.id,
+                                            attributeOptionCombo: found.id,
+                                            period,
+                                            orgUnit
+                                        }]
+                                    });
+                                }
+                            })
+                        }
+                    }
+                });
+            }
+        });
+    } else if (templateType === '4') {
+        let periodMissing = false;
+        let valueMissing = false;
+        let orgUnitMissing = false;
+        rows.forEach(i => {
+            const rowData = categoryCombo.categories.map(category => {
+                const optionCell = category.mapping.value + i;
+                const optionValue = data[optionCell];
+                return optionValue ? optionValue.v : undefined;
+            });
+
+            const found = findAttributeCombo(dataSet, rowData, false);
+            if (found) {
+                _.forOwn(cell2, v => {
+                    const oCell = orgUnitColumn.value + i;
+                    const pCell = periodColumn.value + i;
+                    const vCell = v.value.column + i;
+                    const ouVal = data[oCell];
+                    const periodVal = data[pCell];
+                    const ou = ouVal ? ouVal['v'].toLowerCase() : '';
+                    const period = periodVal ? periodVal['v'] : null;
+                    const val = data[vCell];
+                    const value = val ? val.v : null;
+
+                    const orgUnit = dataSetUnits[ou];
+                    if (orgUnit && value && period) {
+                        dataValues = [...dataValues, {
+                            orgUnit,
+                            period,
+                            value,
+                            dataElement: v.value.dataElement,
+                            attributeOptionCombo: found.id,
+                            categoryOptionCombo: v.value.categoryOptionCombo
+                        }];
+                    } else {
+                        if (!orgUnit) {
+                            orgUnitMissing = true;
+                        }
+                        if (!period) {
+                            periodMissing = true;
+                        }
+
+                        if (!value) {
+                            valueMissing = true;
+                        }
+                    }
+
+                });
+            }
+        });
+
+        if (orgUnitMissing) {
+            // NotificationManager.warning(`Some rows are missing organisation units, will be ignored`);
+        }
+        if (periodMissing) {
+            // NotificationManager.warning(`Some rows are missing periods, will be ignored`);
+        }
+
+        if (valueMissing) {
+            // NotificationManager.warning(`Some rows are missing values, will be ignored`);
+        }
+    }
+    dataValues = dataValues.filter(dv => {
+        return dv.orgUnit && dv.period
+    });
+
+    // if (dataValues.length > 0) {
+    //     return alasql('SELECT orgUnit,dataElement,attributeOptionCombo,categoryOptionCombo,period,SUM(`value`) AS `value` FROM ? GROUP BY orgUnit,dataElement,attributeOptionCombo,categoryOptionCombo,period', [dataValues]);
+    // }
+
+
+    return dataValues
+
+};
+
+export const programUniqueColumn = (program) => {
+    const unique = program.programTrackedEntityAttributes.filter(a => {
+        return a.trackedEntityAttribute.unique && a.column;
+    });
+
+    if (unique.length > 0) {
+        return unique[0]['column']['value'];
+    }
+
+    return null;
+};
+
+export const programUniqueAttribute = (program) => {
+    const unique = program.programTrackedEntityAttributes.filter(a => {
+        return a.trackedEntityAttribute.unique;
+    });
+
+    if (unique.length > 0) {
+        return unique[0]['trackedEntityAttribute']['id'];
+    }
+
+    return null;
+};
+
+
+export const validText = (dataType, value) => {
+    switch (dataType) {
+        case 'TEXT':
+        case 'LONG_TEXT':
+            return value;
+        case 'NUMBER':
+            return !isNaN(value);
+        case 'EMAIL':
+            const re = /\S+@\S+\.\S+/;
+            return re.test(String(value).toLowerCase());
+        case 'BOOLEAN':
+            return value === false || value === true;
+        case 'TRUE_ONLY':
+            return value + '' === true + '';
+        case 'PERCENTAGE':
+            return value >= 0 && value <= 100;
+        case 'INTEGER':
+            return !isNaN(value) && !isNaN(parseInt(value, 10));
+        case 'DATE':
+        case 'DATETIME':
+        case 'TIME':
+            return moment(value).isValid();
+        case 'UNIT_INTERVAL':
+            return value >= 0 && value <= 1;
+        case 'INTEGER_NEGATIVE':
+            return Number.isInteger(value) && value >= 0;
+        case 'NEGATIVE_INTEGER':
+            return Number.isInteger(value) && value < 0;
+        case 'INTEGER_ZERO_OR_POSITIVE':
+        case 'AGE':
+            return Number.isInteger(value) && value >= 0;
+        default:
+            return true
+    }
+};
+
+export const validateValue = (dataType, value, optionSet) => {
+    if (optionSet) {
+        const options = optionSet.options.map(o => {
+            return {
+                code: o.code,
+                value: o.value
+            }
+        });
+        const coded = _.find(options, o => {
+            return value + '' === o.code + '' || value + '' === o.value + '';
+        });
+        if (coded !== undefined && coded !== null) {
+            return coded.code;
+        }
+    } else if (validText(dataType, value)) {
+        if (dataType === 'DATETIME') {
+            return moment(value).format('YYYY-MM-DD HH:mm:ss');
+        } else if (dataType === 'DATE') {
+            return moment(value).format('YYYY-MM-DD');
+        } else if (dataType === 'TIME') {
+            return moment(value).format('HH:mm');
+        }
+        return value;
+        /*const numeric = /^(-?0|-?[1-9]\d*)(\.\d+)?$/;
+        const int = /^(0|-?[1-9]\d*)$/;
+        const posInt = /^[1-9]\d*$/;
+        const posOrZero = /(^0$)|(^[1-9]\d*$)/;
+        const neg = /^-[1-9]\d*$/;*/
+    }
+    return null;
+};
+
+export const searchOrgUnit = (val, orgUnitStrategy, organisationUnits) => {
+    switch (orgUnitStrategy.value) {
+        case 'uid':
+            return _.find(organisationUnits, {
+                id: val
+            });
+        case 'code':
+            return _.find(organisationUnits, {
+                code: val
+            });
+        case 'name':
+            return _.find(organisationUnits, {
+                name: val
+            });
+        case 'auto':
+            const s1 = _.find(organisationUnits, {
+                id: val
+            });
+            const s2 = _.find(organisationUnits, {
+                code: val
+            });
+            const s3 = _.find(organisationUnits, {
+                name: val
+            });
+            if (s1 !== undefined) {
+                return s1;
+            } else if (s2 !== undefined) {
+                return s2;
+            } else if (s3 !== undefined) {
+                return s3;
+            } else {
+                return undefined;
+            }
+        default:
+            return undefined;
+    }
+};
+
+export const getLocation = (href) => {
+    const match = href.match(/^(https?:)\/\/(([^:/?#]*)(?::([0-9]+))?)([/]?[^?#]*)(\?[^#]*|)(#.*|)$/);
+    return match && {
+        href: href,
+        protocol: match[1],
+        host: match[2],
+        hostname: match[3],
+        port: match[4],
+        pathname: match[5],
+        search: match[6],
+        hash: match[7]
+    }
+};
+
+export const removeDuplicates = (evs, stageEventFilters) => {
+    if (stageEventFilters && stageEventFilters.elements && stageEventFilters.event) {
+        evs = _.uniqBy(evs, v => {
+            const filteredAndSame = stageEventFilters.elements.map(se => {
+                const foundPrevious = _.filter(v.dataValues, {
+                    dataElement: se
+                });
+                if (foundPrevious.length > 0) {
+                    const exists = foundPrevious[0].value;
+                    return {
+                        exists
+                    };
+                } else {
+                    return {
+                        exists: false
+                    }
+                }
+            });
+
+            if (_.some(filteredAndSame, {
+                'exists': false
+            })) {
+                return v.event;
+            } else {
+                return JSON.stringify([v.eventDate, filteredAndSame])
+            }
+        });
+
+    } else if (stageEventFilters && stageEventFilters.elements) {
+
+        evs = _.uniqBy(evs, v => {
+            const filteredAndSame = stageEventFilters.elements.map(se => {
+                const foundPrevious = _.filter(v.dataValues, {
+                    dataElement: se
+                });
+                if (foundPrevious.length > 0) {
+                    const exists = foundPrevious[0].value;
+                    return {
+                        exists
+                    };
+                } else {
+                    return {
+                        exists: false
+                    }
+                }
+            });
+
+            if (_.some(filteredAndSame, {
+                'exists': false
+            })) {
+                return v.event;
+            } else {
+                return JSON.stringify([filteredAndSame])
+            }
+        });
+
+    } else if (stageEventFilters && stageEventFilters.event) {
+        evs = _.uniqBy(evs, v => {
+            return v.eventDate;
+        });
+    }
+
+    return evs;
+};
+
+export const searchEvent = (enrollmentEvents, stageEventFilters, stage, e) => {
+    return _.findIndex(enrollmentEvents, item => {
+        if (!stageEventFilters) {
+            return false
+        } else if (stageEventFilters.elements && stageEventFilters.event) {
+            const filteredAndSame = stageEventFilters.elements.map(se => {
+                const foundPrevious = _.filter(item.dataValues, {
+                    dataElement: se
+                });
+                const foundCurrent = _.filter(e.dataValues, {
+                    dataElement: se
+                });
+                if (foundCurrent.length > 0 && foundPrevious.length > 0) {
+                    const exists = foundPrevious[0].value === foundCurrent[0].value;
+                    return {
+                        exists
+                    };
+                } else {
+                    return {
+                        exists: false
+                    }
+                }
+            });
+            return item.programStage === stage &&
+                moment(item.eventDate, 'YYYY-MM-DD').format('YYYY-MM-DD') ===
+                moment(e.eventDate, 'YYYY-MM-DD').format('YYYY-MM-DD') &&
+                _.every(filteredAndSame, 'exists');
+        } else if (stageEventFilters.elements) {
+            const filteredAndSame = stageEventFilters.elements.map(se => {
+                const foundPrevious = _.filter(item.dataValues, {
+                    dataElement: se
+                });
+                const foundCurrent = _.filter(e.dataValues, {
+                    dataElement: se
+                });
+                if (foundCurrent.length > 0 && foundPrevious > 0) {
+                    return {
+                        exists: foundPrevious[0].value === foundCurrent[0].value
+                    };
+                } else {
+                    return {
+                        exists: false
+                    }
+                }
+            });
+
+            return item.programStage === stage && _.every(filteredAndSame, 'exists')
+        } else if (stageEventFilters.event) {
+            return item.programStage === stage &&
+                moment(item.eventDate, 'YYYY-MM-DD').format('YYYY-MM-DD') ===
+                moment(e.eventDate, 'YYYY-MM-DD').format('YYYY-MM-DD')
+        }
+    });
+};
+
+export const isTracker = (program) => {
+    return program.programType === 'WITH_REGISTRATION';
+};
+
+export const groupEntities = (attribute, trackedEntityInstances,) => {
+    const entities = trackedEntityInstances.map(e => {
+        const uniqueAttribute = _.find(e.attributes, {
+            attribute
+        });
+        const val = uniqueAttribute ? uniqueAttribute['value'] : null;
+        return {
+            ...e,
+            ..._.fromPairs([
+                [attribute, val]
+            ])
+        }
+    });
+    return _.groupBy(entities, attribute);
+};
+
+
+export const processProgramData = (data, program, uniqueColumn, instances, isTracker) => {
+
+    let eventsUpdate = [];
+    let trackedEntityInstancesUpdate = [];
+
+    let newEvents = [];
+    let newEnrollments = [];
+    let newTrackedEntityInstances = [];
+
+    let duplicates = [];
+    let conflicts = [];
+    let errors = [];
+
+    const {
+        id,
+        programStages,
+        dataSource,
+        programTrackedEntityAttributes,
+        incidentDateColumn,
+        enrollmentDateColumn,
+        trackedEntityType,
+        trackedEntity,
+        updateEntities,
+        createEntities,
+        createNewEnrollments,
+        orgUnitStrategy,
+        orgUnitColumn
+    } = program;
+
+
+    if (uniqueColumn) {
+        data = data.filter(d => {
+            return d[uniqueColumn] !== null && d[uniqueColumn] !== undefined;
+        });
+        let clients = _.groupBy(data, uniqueColumn);
+        let newClients = [];
+        _.forOwn(clients, (data, client) => {
+            const previous = instances[client] || [];
+            newClients = [...newClients, {
+                client,
+                data,
+                previous
+            }];
+        });
+        data = newClients;
+    } else if (data && data.length > 0) {
+        data = data.map((data, i) => {
+            return {
+                data: [data],
+                client: i + 1,
+                previous: []
+            };
+        });
+    }
+
+    if (data && data.length > 0) {
+        data.forEach(client => {
+            let events = [];
+            let allAttributes = [];
+            let currentData = client.data;
+            let enrollmentDates = [];
+            let orgUnits = [];
+            let identifierElements = {};
+            currentData.forEach(d => {
+                programStages.forEach(stage => {
+                    let dataValues = [];
+                    let eventDate;
+                    if (stage.eventDateColumn && (stage.createNewEvents || stage.updateEvents) && dataSource === 2) {
+                        const date = moment(d[stage.eventDateColumn.value]);
+                        if (date.isValid()) {
+                            eventDate = date.format('YYYY-MM-DD');
+                        }
+                    } else if (stage.eventDateColumn && (stage.createNewEvents || stage.updateEvents)) {
+                        const date = moment(d[stage.eventDateColumn.value], 'YYYY-MM-DD');
+                        if (date.isValid()) {
+                            eventDate = date.format('YYYY-MM-DD');
+                        }
+                    }
+
+                    const mapped = stage.programStageDataElements.filter(e => {
+                        return e.column && e.column.value
+                    });
+
+                    identifierElements[stage.id] = {
+                        elements: mapped.filter(e => {
+                            return e.dataElement.identifiesEvent;
+                        }).map(e => e.dataElement.id),
+                        event: stage.eventDateIdentifiesEvent
+                    };
+                    // Coordinates
+                    let coordinate = null;
+                    if (stage.latitudeColumn && stage.longitudeColumn) {
+                        coordinate = {
+                            latitude: d[stage.latitudeColumn.value],
+                            longitude: d[stage.longitudeColumn.value]
+                        };
+                    }
+                    if (eventDate && mapped.length > 0) {
+                        mapped.forEach(e => {
+                            const value = d[e.column.value];
+                            const type = e.dataElement.valueType;
+                            const optionsSet = e.dataElement.optionSet;
+                            const validatedValue = validateValue(type, value, optionsSet);
+
+                            if (value !== '' && validatedValue !== null) {
+                                dataValues = [...dataValues, {
+                                    dataElement: e.dataElement.id,
+                                    value: validatedValue
+                                }];
+                            } else if (value !== undefined) {
+                                conflicts = [...conflicts, {
+                                    error: optionsSet === null ? 'Invalid value ' + value + ' for value type ' + type : 'Invalid value: ' + value + ', expected: ' + _.map(optionsSet.options, o => {
+                                        return o.code
+                                    }).join(','),
+                                    row: client.client,
+                                    column: e.column.value
+                                }]
+                            }
+                        });
+
+                        let event = {
+                            dataValues,
+                            eventDate,
+                            programStage: stage.id,
+                            program: id,
+                            event: generateUid()
+                        };
+
+                        if (coordinate) {
+                            event = {
+                                ...event,
+                                coordinate
+                            }
+                        }
+
+                        if (stage.completeEvents) {
+                            event = {
+                                ...event,
+                                ...{
+                                    status: 'COMPLETED',
+                                    completedDate: event['eventDate']
+                                }
+                            }
+                        }
+
+                        events = [...events, event];
+                    }
+                });
+
+                const mappedAttributes = programTrackedEntityAttributes.filter(a => {
+                    return a.column && a.column.value
+                });
+
+                let attributes = [];
+
+                mappedAttributes.forEach(a => {
+                    const value = d[a.column.value];
+                    const type = a.valueType;
+                    const optionsSet = a.trackedEntityAttribute.optionSet;
+                    const validatedValue = validateValue(type, value, optionsSet);
+
+                    if (value !== '' && validatedValue !== null) {
+                        attributes = [...attributes, {
+                            attribute: a.trackedEntityAttribute.id,
+                            value: validatedValue
+                        }]
+                    } else if (value !== undefined) {
+                        conflicts = [...conflicts, {
+                            error: !optionsSet ? 'Invalid value ' + value + ' for value type ' + type : 'Invalid value ' + value + ' choose from options: ' +
+                                _.map(optionsSet.options, o => o.code).join(','),
+                            row: client.client,
+                            column: a.column.value
+                        }]
+                    }
+
+                });
+
+                if (attributes.length > 0) {
+                    allAttributes = [...allAttributes, attributes];
+                }
+
+                if (isTracker && enrollmentDateColumn && incidentDateColumn) {
+                    const enrollmentDate = moment(d[enrollmentDateColumn.value], 'YYYY-MM-DD');
+                    const incidentDate = moment(d[incidentDateColumn.value], 'YYYY-MM-DD');
+
+                    if (enrollmentDate.isValid() && incidentDate.isValid()) {
+                        enrollmentDates = [...enrollmentDates, {
+                            enrollmentDate: enrollmentDate.format('YYYY-MM-DD'),
+                            incidentDate: incidentDate.format('YYYY-MM-DD')
+                        }]
+                    }
+                }
+
+                if (orgUnitColumn !== '') {
+                    orgUnits = [...orgUnits, d[orgUnitColumn.value]]
+                }
+            });
+            let groupedEvents = _.groupBy(events, 'programStage');
+
+            if (client.previous.length > 1) {
+                duplicates = [...duplicates, {identifier: client.client}]
+            } else if (client.previous.length === 1) {
+                client.previous.forEach(p => {
+                    let enrollments = p['enrollments'];
+                    if (updateEntities) {
+                        const nAttributes = _.differenceWith(allAttributes[0], p['attributes'], (a, b) => {
+                            return a.attribute === b.attribute && a.value + '' === b.value + '';
+                        });
+                        if (nAttributes.length > 0) {
+                            const mergedAttributes = _.unionBy(allAttributes[0], p['attributes'], 'attribute');
+                            let tei;
+
+                            if (trackedEntityType) {
+                                tei = {
+                                    ..._.pick(p, ['orgUnit', 'trackedEntityInstance', 'trackedEntityType']),
+                                    attributes: mergedAttributes
+                                };
+                            } else if (trackedEntity) {
+                                tei = {
+                                    ..._.pick(p, ['orgUnit', 'trackedEntityInstance', 'trackedEntity']),
+                                    attributes: mergedAttributes
+                                };
+                            }
+                            trackedEntityInstancesUpdate = [...trackedEntityInstancesUpdate, tei];
+                        }
+                    }
+                    events = events.map(e => {
+                        return {
+                            ...e,
+                            trackedEntityInstance: p['trackedEntityInstance'],
+                            orgUnit: p['orgUnit']
+                        }
+                    });
+
+                    groupedEvents = _.groupBy(events, 'programStage');
+                    const enrollmentIndex = _.findIndex(enrollments, {
+                        program: id
+                    });
+                    if (enrollmentIndex === -1 && createNewEnrollments && enrollmentDates.length > 0) {
+                        let enroll = {
+                            program: id,
+                            orgUnit: p['orgUnit'],
+                            trackedEntityInstance: p['trackedEntityInstance'],
+                            ...enrollmentDates[0]
+                        };
+                        newEnrollments = [...newEnrollments, enroll];
+                        _.forOwn(groupedEvents, (evs, stage) => {
+                            const stageEventFilters = identifierElements[stage];
+                            const stageInfo = _.find(programStages, {
+                                id: stage
+                            });
+                            const {
+                                repeatable,
+                                createNewEvents,
+                            } = stageInfo;
+
+                            evs = removeDuplicates(evs, stageEventFilters);
+                            if (createNewEvents) {
+                                if (!repeatable) {
+                                    const ev = _.maxBy(evs, 'eventDate');
+                                    if (ev.dataValues.length > 0) {
+                                        newEvents = [...newEvents, ev];
+                                    }
+                                } else {
+                                    newEvents = [...newEvents, ...evs];
+                                }
+                            }
+
+                        });
+
+                        enrollments = [...enrollments, enroll];
+                        p = {
+                            ...p,
+                            enrollments
+                        }
+                    } else if (enrollmentIndex === -1 && enrollmentDates.length === 0) {
+                        console.log('Ignoring new enrollments');
+                    } else if (enrollmentIndex !== -1) {
+                        let enrollment = enrollments[enrollmentIndex];
+                        let enrollmentEvents = enrollment['events'];
+                        _.forOwn(groupedEvents, (evs, stage) => {
+                            const stageInfo = _.find(programStages, {
+                                id: stage
+                            });
+                            const {
+                                repeatable,
+                                updateEvents,
+                                createNewEvents
+                            } = stageInfo;
+
+                            const stageEventFilters = identifierElements[stage];
+
+                            evs = removeDuplicates(evs, stageEventFilters);
+
+                            if (repeatable) {
+                                evs.forEach(e => {
+                                    const eventIndex = searchEvent(enrollmentEvents, stageEventFilters, stage, e);
+                                    if (eventIndex !== -1 && updateEvents) {
+                                        const stageEvent = enrollmentEvents[eventIndex];
+                                        const differingElements = _.differenceWith(e['dataValues'], stageEvent['dataValues'], (a, b) => {
+                                            return a.dataElement === b.dataElement && a.value + '' === b.value + '';
+                                        });
+                                        if (differingElements.length > 0) {
+                                            const mergedEvent = {
+                                                ...stageEvent,
+                                                dataValues: differingElements
+                                            };
+                                            eventsUpdate = [...eventsUpdate, mergedEvent];
+                                        }
+                                    } else if (eventIndex === -1 && createNewEvents) {
+                                        newEvents = [...newEvents, e];
+                                    }
+                                });
+                            } else {
+                                let foundEvent = _.find(enrollmentEvents, {
+                                    programStage: stage
+                                });
+                                let max = _.maxBy(evs, 'eventDate');
+                                if (foundEvent && updateEvents) {
+                                    const differingElements = _.differenceWith(max['dataValues'], foundEvent['dataValues'], (a, b) => {
+                                        return a.dataElement === b.dataElement && a.value + '' === b.value + '';
+                                    });
+                                    if (differingElements.length > 0) {
+                                        const mergedEvent = {
+                                            ...foundEvent,
+                                            dataValues: differingElements
+                                        };
+                                        eventsUpdate = [...eventsUpdate, mergedEvent];
+                                    }
+                                } else if (!foundEvent && createNewEvents) {
+                                    newEvents = [...newEvents, max];
+                                }
+                            }
+                        });
+                    }
+                });
+            } else {
+                orgUnits = _.uniq(orgUnits);
+                let orgUnit;
+                if (orgUnits.length > 1) {
+                    errors = [...errors, {
+                        error: 'Entity belongs to more than one organisation unit',
+                        row: client.client
+                    }]
+                } else if (orgUnits.length === 1) {
+                    orgUnit = searchOrgUnit(orgUnits[0]);
+                    if (orgUnit) {
+                        if (isTracker) {
+                            const trackedEntityInstance = generateUid();
+
+                            if (createEntities) {
+                                let tei = {
+                                    orgUnit: orgUnit.id,
+                                    attributes: allAttributes[0],
+                                    trackedEntityInstance
+                                };
+
+                                if (trackedEntityType) {
+                                    tei = {
+                                        ...tei,
+                                        trackedEntityType: trackedEntityType.id
+                                    }
+                                } else if (trackedEntity && trackedEntity.id) {
+                                    tei = {
+                                        ...tei,
+                                        trackedEntity: trackedEntity.id
+                                    }
+                                }
+                                newTrackedEntityInstances = [...newTrackedEntityInstances, tei];
+                            }
+
+                            if (createNewEnrollments) {
+
+                                let enrollment = {
+                                    orgUnit: orgUnit.id,
+                                    program: id,
+                                    trackedEntityInstance,
+                                    ...enrollmentDates[0],
+                                    enrollment: generateUid()
+                                };
+
+                                newEnrollments = [...newEnrollments, enrollment];
+
+                            }
+
+                            _.forOwn(groupedEvents, (evs, stage) => {
+                                const stageEventFilters = identifierElements[stage];
+                                const stageInfo = _.find(programStages, {
+                                    id: stage
+                                });
+                                const {
+                                    repeatable,
+                                    createNewEvents
+                                } = stageInfo;
+                                evs = evs.map(e => {
+                                    return {
+                                        ...e,
+                                        orgUnit: orgUnit.id,
+                                        event: generateUid(),
+                                        trackedEntityInstance
+                                    }
+                                });
+
+                                evs = removeDuplicates(evs, stageEventFilters);
+
+                                if (createNewEvents) {
+                                    if (!repeatable) {
+                                        newEvents = [...newEvents, _.maxBy(evs, 'eventDate')];
+                                    } else {
+                                        newEvents = [...newEvents, ...evs]
+                                    }
+                                }
+                            });
+                        } else if (!isTracker) {
+                            events = events.map(e => {
+                                return {
+                                    ...e,
+                                    orgUnit: orgUnit.id
+                                }
+                            });
+                            newEvents = [...newEvents, ...events];
+                        }
+                    } else {
+                        errors = [...errors, {
+                            error: 'Organisation unit ' + orgUnits[0] + ' not found using strategy ' +
+                                orgUnitStrategy.value,
+                            row: client.client
+                        }]
+                    }
+                } else if (orgUnits.length === 0) {
+                    errors = [...errors, {
+                        error: 'Organisation unit missing',
+                        row: client.client
+                    }]
+                }
+            }
+        });
+    }
+
+    return {
+        newTrackedEntityInstances,
+        newEnrollments,
+        newEvents,
+        trackedEntityInstancesUpdate,
+        eventsUpdate,
+        conflicts,
+        duplicates,
+        errors
+    }
 };
